requirements.txt
requests
beautifulsoup4
PyYAML
boto3